Add html files and images
Here, I have created a index.html which is basically a homepage of my
website. In the index.html, there are two links which redirects the page
to the html files of recipes directory.

In recipes directory, there are two files and each file consists of
heading tags and list tag with images. The required image for the
website stored in images directory.